refactor: Moves pdf_cache_service to services directory
Relocates the pdf_cache_service from the core directory to the services directory.

This change improves the project's structure by grouping services in a dedicated directory.
