+++ frontend/lib/features/player/services/pdf_cache_service.dart
